👌 Rehabilitate getMutationObserverConstructor()
