+++ packages/rum-core/src/index.ts
@@ -14,4 +14,4 @@ export { startRum } from './boot/rum'
 export { LifeCycle, LifeCycleEventType } from './domain/lifeCycle'
 export { ParentContexts } from './domain/parentContexts'
 export { RumSession } from './domain/rumSession'
-export { getMutationObserverConstructor } from './browser/domMutationCollection'
+export { getMutationObserverConstructor } from './browser/domMutationObservable'
